Split wigetutil tests into files
